Accept gzip encoding of GraphQL response
To increase the amount of information that can be retrieved from a
server the response must be allowed to be compressed. This commit
enables the Accept-Encoding header to allow gzip response. Currently
there is no checking that the response is actually in the gzip
encoding so if gzip encoding is unsupported on the server the request
will fail. This checking will need to be added in a future commit.